Add AccountNotificationStatus

+++ src/main/kotlin/me/ewan/cellit/domain/account/vo/query/AccountNotificationQuery.kt
@@ -16,7 +16,7 @@ data class AccountNotificationQuery(
 
         var createDate: String? = null,
 
-        override var offset: Int? = null,
+        override var offset: Long? = null,
 
-        override var limit: Int? = null
+        override var limit: Long? = null
 ) : Query()
