+++ src/main/kotlin/me/ewan/cellit/global/common/Query.kt
@@ -1,6 +1,6 @@
 package me.ewan.cellit.global.common
 
 abstract class Query(
-        open var offset: Int? = null,
-        open var limit: Int? = null
+        open var offset: Long? = null,
+        open var limit: Long? = null
 )
